replaced old network monitor with new
