+++ packages/network-monitor/src/screens/Network/sol_stake.tsx
