+++ packages/network-monitor/src/screens/Stablecoins/defi.tsx
