+++ packages/network-monitor/src/screens/Stablecoins/index.tsx
