--- packages/solana-monitor/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "network-monitor",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "xnft build",
-    "start": "xnft watch",
-    "dev": "xnft dev"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@coral-xyz/xnft-cli": "^0.0.1",
-    "@parcel/config-default": "^2.7.0"
-  },
-  "dependencies": {
-    "react": "^17.0.2",
-    "react-xnft": "latest",
-    "swr": "^1.3.0"
-  }
-}
