--- packages/solana-monitor/src/index.tsx
@@ -1,8 +1,0 @@
-import ReactXnft, { AnchorDom } from "react-xnft";
-import { App } from "./app";
-
-ReactXnft.render(
-  <AnchorDom>
-    <App />
-  </AnchorDom>
-);
